+++ lib/src/transitive_use/transitive_use_checker.dart
@@ -6,56 +6,50 @@ import 'package:dart_dependency_checker/src/util/dart_files.dart';
 import 'package:dart_dependency_checker/src/util/iterable_ext.dart';
 import 'package:dart_dependency_checker/src/util/pubspec_yaml_loader.dart';
 import 'package:dart_dependency_checker/src/util/yaml_map_ext.dart';
-import 'package:yaml/yaml.dart';
 
 /// Checks direct use of undeclared/transitive dependencies.
 class TransitiveUseChecker
     extends Checker<TransitiveUseParams, TransitiveUseResults> {
   const TransitiveUseChecker(super.params);
 
-  /// TODO(albert): fix this...
-  /// - ignore existing main dependencies when checking dev dependencies
-  /// - ignore own package
-  // ================ DDC ================
-  // Path: /Users/albert.kaiser/Dev/klar/klar-mobile-app/klar_user_contacts_repository/pubspec.yaml
-  // Message: Found undeclared/transitive packages
-  // Dependencies:
-  //   - collection
-  //   - klar_user_contacts_repository      !!!this is own package!!!
-  //   - klar_data_result
-  // Dev Dependencies:
-  //   - klar_api_client                    !!!this is declared in main!!!
-  //   - klar_user_contacts_repository      !!!this is own package!!!
-  //   - klar_data_result
-  //   - mocktail
-  ///
   @override
   TransitiveUseResults check() {
     final pubspecYaml = PubspecYamlLoader.from(params.path);
+    final ownReference = pubspecYaml.name;
+
+    final declaredMainDependencies =
+        pubspecYaml.packages(DependencyType.mainDependencies);
 
     return TransitiveUseResults(
       mainDependencies: _find(
-        pubspecYaml,
         DependencyType.mainDependencies,
-        params.mainIgnores,
+        {
+          ...params.mainIgnores,
+          ownReference,
+        },
+        (_) => declaredMainDependencies,
       ),
       devDependencies: _find(
-        pubspecYaml,
         DependencyType.devDependencies,
-        params.devIgnores,
+        {
+          ...params.devIgnores,
+          ...declaredMainDependencies,
+          ownReference,
+        },
+        (dependencyType) => pubspecYaml.packages(dependencyType),
       ),
     );
   }
 
   Set<String> _find(
-    YamlMap pubspecYaml,
     DependencyType dependencyType,
-    Set<String> ignores,
+    Set<String?> ignores,
+    Set<String> Function(DependencyType) declaredDependencies,
   ) =>
       DartFiles.from(params.path, dependencyType)
           .map((file) => DartFiles.packages(file))
           .expand((packages) => packages)
           .unmodifiable
-          .difference(pubspecYaml.packages(dependencyType))
-          .difference(ignores);
+          .difference(declaredDependencies(dependencyType))
+          .difference(ignores..nonNulls);
 }