+++ lib/src/util/yaml_map_ext.dart
@@ -3,6 +3,8 @@ import 'package:dart_dependency_checker/src/util/iterable_ext.dart';
 import 'package:yaml/yaml.dart';
 
 extension YamlMapExt on YamlMap {
+  String? get name => this['name'] as String?;
+
   Set<String> packages(DependencyType dependencyType) {
     final nodeValue = nodes[dependencyType.yamlNode]?.value as YamlMap?;
     return nodeValue?.keys.map((e) => e as String).unmodifiable ?? const {};