+++ test/resources/samples/in_main_but_missing_in_dev/lib/main.dart
@@ -1,0 +1,2 @@
+// ignore_for_file: depend_on_referenced_packages, unused_import
+import 'package:args/args.dart';