+++ test/resources/samples/in_main_but_missing_in_dev/test/test.dart
@@ -1,0 +1,3 @@
+// ignore_for_file: depend_on_referenced_packages, unused_import
+import 'package:args/args.dart';
+import 'package:test/test.dart';