+++ test/resources/samples/own_reference/lib/main.dart
@@ -1,0 +1,3 @@
+// ignore_for_file: depend_on_referenced_packages, unused_import
+import 'package:args/args.dart';
+import 'package:dart_dependency_checker/dart_dependency_checker.dart';