+++ test/resources/samples/own_reference/test/test.dart
@@ -1,0 +1,3 @@
+// ignore_for_file: unused_import
+import 'package:dart_dependency_checker/dart_dependency_checker.dart';
+import 'package:test/test.dart';