+++ test/src/_paths.dart
@@ -2,8 +2,11 @@ const //
     allSourcesDirsPath = 'test/resources/samples/all_sources_dirs',
     allSourcesDirsMultiPath = 'test/resources/samples/all_sources_dirs_multi',
     emptyYamlPath = 'test/resources/samples/empty_yaml',
+    inMainButMissingInDev = 'test/resources/samples/in_main_but_missing_in_dev',
     meantForFixingPath = 'test/resources/samples/meant_for_fixing',
+    missingNameNodePath = 'test/resources/samples/missing_name_node',
     noDependenciesPath = 'test/resources/samples/no_dependencies',
     noSourcesDirsPath = 'test/resources/samples/no_sources_dirs',
+    ownReferencePath = 'test/resources/samples/own_reference',
     yamlAndYmlFilesPath = 'test/resources/samples/yaml_and_yml_files',
     ymlFilePath = 'test/resources/samples/yml_file';